add HttpRequestLogger middleware  - Logs the full HTTP request to text files for debugging purposes
Note: I wanted to be able to enable this and set filters from the settings file, hower, when I tried to do that, I got errors because the settings object wasn't initialized for the request yet.  So I tried moving the registry manager middleware above this middleware, but that I remembered that the registry objects get initilized in pysmvt.application:WSGIApplication().

So, initilizing the registry settings AND objects needs to happen as one of the first middlewares so that other middleware has access to the settings object.