dispatch testing & new "minimal testing" helpers
up until now, testing controllers or views has always taken place
by having an application on the file system.  That can slow things
down when testing because it requires a view, route, and test to
be created.  With the new helpers, we can easily control what
happens when the view is executed, without needing so much
boilerplate.