**BC BREAK**: mock_smtp => mockmail
mockmail() mocks objects at a higher level than mock_smtp does.
The reason for changing this is that the email messages is
chopped up when getting sent to smtplib.  It is easier to
test if we can get it earlier, before it gets chopped up.

The order of output for mockmail is a bit different, see
test_email.py for an example of what to look for.