add new config option http_exception_handling; defaults to True
Can be set to False and it will result in HTTPExceptions being
re-raised instead of caught and turned into responses.  This can
help in debugging when you are getting 3xx,4xx,5xx status code responses
but are not sure why.  Set this to False in your testing settings
profile, re-run the test, and the exception will come all the way through.